Hide footer bar
In order to get the most possible clean newTab I wrote 0 in number of
tiles. But it was not enough, the footer bar should disappear. So I
investigate a bit and I discovered the github project. I started working
in my idea and... I can hide the footer bar now.

So, my fork has a new option to hide the footer.